requirements.txt
apscheduler==3.6.3
Cython==0.29.21
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy>=1.22.4
pandas==1.4.2
Pillow==8.0.1
pydub==0.24.1
requests==2.24.0
requests-oauthlib==1.3.0
tqdm==4.50.0
yapf==0.31.0